backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
transformers==4.37.2
torch
pydantic==2.6.1
PyPDF2==3.0.1
python-multipart==0.0.9
google-cloud-vision==3.5.0 